# Conflicts:
#	modules/platforms/dotnet/Apache.Ignite/Internal/Table/RecordView.cs

+++ modules/storage-page-memory/src/main/java/org/apache/ignite/internal/storage/pagememory/VolatilePageMemoryTableStorage.java
@@ -182,9 +182,14 @@ public class VolatilePageMemoryTableStorage extends AbstractPageMemoryTableStora
     CompletableFuture<Void> destroyMvPartitionStorage(AbstractPageMemoryMvPartitionStorage mvPartitionStorage) {
         VolatilePageMemoryMvPartitionStorage volatilePartitionStorage = (VolatilePageMemoryMvPartitionStorage) mvPartitionStorage;
 
-        volatilePartitionStorage.transitionToDestroyingState();
+        boolean transitioned = volatilePartitionStorage.transitionToDestroyedState();
+        if (!transitioned) {
+            // Someone has already started destruction (in which case we don't need to do anything), or closed the storage
+            // (which means the node is going down, so the volatile storage does not need to be destroyed).
+            return nullCompletedFuture();
+        }
 
         return volatilePartitionStorage.destroyStructures()
-                .whenComplete((v, e) -> volatilePartitionStorage.close());
+                .whenComplete((v, e) -> volatilePartitionStorage.closeResources());
     }
 }
